Safari-611.1.10/LayoutTests/accessibility/radio-button-title-label-expected.txt

This test checks that radio buttons expose title ui elements correctly under a variety of cirmcumstances. In general, the should NOT disappear and should be the titleUIElement for the checkbox.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS radio1.title is 'AXTitle: '
PASS titleUIElement.isEqual(accessibilityController.accessibleElementById('label1')) is true
PASS radio2.description is 'AXDescription: LABEL2a'
PASS radio2.title is 'AXTitle: '
PASS !titleUIElement || titleUIElement.title == 'AXTitle: ' is true
PASS radio3.description is 'AXDescription: radio3'
PASS radio3.title is 'AXTitle: '
PASS !titleUIElement || titleUIElement.title == 'AXTitle: ' is true
PASS successfullyParsed is true

TEST COMPLETE
Test
LABEL  Test
LABEL2
LABEL2a
Test
LABEL3
